requirements.txt
acres==0.5.0
annotated-types==0.6.0
anyio==4.10.0
asttokens==3.0.0
boltons==24.1.0
brotli-python==1.0.9
cachetools==5.5.2
cbor2==5.7.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
crcmod==1.7
cryptography==44.0.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.2.1
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
fonttools==4.59.1
frozendict==2.4.2
fsspec==2025.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.7
imageio==2.37.0
importlib-metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==2.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-text-splitters==0.3.9
langdetect==1.0.9
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.14
lazy-loader==0.4
looseversion==1.3.0
lxml==6.0.0
markdown-it-py==2.2.0
markupsafe==3.0.2
matplotlib==3.10.5
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
orjson==3.11.2
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
pathlib==1.0.1
pillow==11.3.0
pip==25.0
platformdirs==4.3.7
pluggy==1.5.0
proto-plus==1.26.1
protobuf==6.31.1
prov==2.1.1
psutil==5.9.0
puremagic==1.29
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.10.3
pydot==4.0.1
pygments==2.19.1
pyparsing==3.2.3
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
pyyaml==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
sarvamai==0.1.17
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
segmentation-models-pytorch==0.5.0
sentence-transformers==5.1.0
setuptools==78.1.1
simplejson==3.20.1
six==1.17.0
smp==3.3.2
sniffio==1.3.1
sqlalchemy==2.0.43
stack_data==0.6.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.19
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traits==7.0.2
transformers==4.55.2
triton==3.4.0
typing-extensions==4.12.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
websockets==15.0.1
wheel==0.45.1
xxhash==3.5.0
zipp==3.23.0
zstandard==0.23.0
